@@ -1,8 +1,9 @@
 package vhost
 
 import (
+	"cmp"
 	"errors"
-	"sort"
+	"slices"
 	"strings"
 	"sync"
 )
@@ -33,6 +34,8 @@ func NewRouters() *Routers {
 }
 
 func (r *Routers) Add(domain, location, httpUser string, payload interface{}) error {
+	domain = strings.ToLower(domain)
+
 	r.mutex.Lock()
 	defer r.mutex.Unlock()
 
@@ -56,7 +59,10 @@ func (r *Routers) Add(domain, location, httpUser string, payload interface{}) er
 		payload:  payload,
 	}
 	vrs = append(vrs, vr)
-	sort.Sort(sort.Reverse(ByLocation(vrs)))
+
+	slices.SortFunc(vrs, func(a, b *Router) int {
+		return -cmp.Compare(a.location, b.location)
+	})
 
 	routersByHTTPUser[httpUser] = vrs
 	r.indexByDomain[domain] = routersByHTTPUser
@@ -64,6 +70,8 @@ func (r *Routers) Add(domain, location, httpUser string, payload interface{}) er
 }
 
 func (r *Routers) Del(domain, location, httpUser string) {
+	domain = strings.ToLower(domain)
+
 	r.mutex.Lock()
 	defer r.mutex.Unlock()
 
@@ -86,6 +94,8 @@ func (r *Routers) Del(domain, location, httpUser string) {
 }
 
 func (r *Routers) Get(host, path, httpUser string) (vr *Router, exist bool) {
+	host = strings.ToLower(host)
+
 	r.mutex.RLock()
 	defer r.mutex.RUnlock()
 
@@ -123,19 +133,4 @@ func (r *Routers) exist(host, path, httpUser string) (route *Router, exist bool)
 		}
 	}
 	return
-}
-
-// sort by location
-type ByLocation []*Router
-
-func (a ByLocation) Len() int {
-	return len(a)
-}
-
-func (a ByLocation) Swap(i, j int) {
-	a[i], a[j] = a[j], a[i]
-}
-
-func (a ByLocation) Less(i, j int) bool {
-	return strings.Compare(a[i].location, a[j].location) < 0
 }
