@@ -27,11 +27,11 @@ func Register(m ServerMetrics) {
 
 type noopServerMetrics struct{}
 
-func (noopServerMetrics) NewClient()                                                      {}
-func (noopServerMetrics) CloseClient()                                                    {}
-func (noopServerMetrics) NewProxy(name string, proxyType string)                          {}
-func (noopServerMetrics) CloseProxy(name string, proxyType string)                        {}
-func (noopServerMetrics) OpenConnection(name string, proxyType string)                    {}
-func (noopServerMetrics) CloseConnection(name string, proxyType string)                   {}
-func (noopServerMetrics) AddTrafficIn(name string, proxyType string, trafficBytes int64)  {}
-func (noopServerMetrics) AddTrafficOut(name string, proxyType string, trafficBytes int64) {}
+func (noopServerMetrics) NewClient()                          {}
+func (noopServerMetrics) CloseClient()                        {}
+func (noopServerMetrics) NewProxy(string, string)             {}
+func (noopServerMetrics) CloseProxy(string, string)           {}
+func (noopServerMetrics) OpenConnection(string, string)       {}
+func (noopServerMetrics) CloseConnection(string, string)      {}
+func (noopServerMetrics) AddTrafficIn(string, string, int64)  {}
+func (noopServerMetrics) AddTrafficOut(string, string, int64) {}
