@@ -9,8 +9,7 @@ import (
 	"strings"
 	"text/template"
 
-	"github.com/onsi/ginkgo"
-	"github.com/onsi/ginkgo/config"
+	"github.com/onsi/ginkgo/v2"
 
 	"github.com/fatedier/frp/test/e2e/mock/server"
 	"github.com/fatedier/frp/test/e2e/pkg/port"
@@ -30,8 +29,8 @@ type Framework struct {
 	// ports used in this framework indexed by port name.
 	usedPorts map[string]int
 
-	// record ports alloced by this framework and release them after each test
-	allocedPorts []int
+	// record ports allocated by this framework and release them after each test
+	allocatedPorts []int
 
 	// portAllocator to alloc port for this test case.
 	portAllocator *port.Allocator
@@ -63,11 +62,12 @@ type Framework struct {
 }
 
 func NewDefaultFramework() *Framework {
+	suiteConfig, _ := ginkgo.GinkgoConfiguration()
 	options := Options{
-		TotalParallelNode: config.GinkgoConfig.ParallelTotal,
-		CurrentNodeIndex:  config.GinkgoConfig.ParallelNode,
-		FromPortIndex:     20000,
-		ToPortIndex:       50000,
+		TotalParallelNode: suiteConfig.ParallelTotal,
+		CurrentNodeIndex:  suiteConfig.ParallelProcess,
+		FromPortIndex:     10000,
+		ToPortIndex:       30000,
 	}
 	return NewFramework(options)
 }
@@ -118,14 +118,14 @@ func (f *Framework) AfterEach() {
 	// stop processor
 	for _, p := range f.serverProcesses {
 		_ = p.Stop()
-		if TestContext.Debug {
+		if TestContext.Debug || ginkgo.CurrentSpecReport().Failed() {
 			fmt.Println(p.ErrorOutput())
 			fmt.Println(p.StdOutput())
 		}
 	}
 	for _, p := range f.clientProcesses {
 		_ = p.Stop()
-		if TestContext.Debug {
+		if TestContext.Debug || ginkgo.CurrentSpecReport().Failed() {
 			fmt.Println(p.ErrorOutput())
 			fmt.Println(p.StdOutput())
 		}
@@ -153,11 +153,11 @@ func (f *Framework) AfterEach() {
 	}
 	f.usedPorts = make(map[string]int)
 
-	// release alloced ports
-	for _, port := range f.allocedPorts {
+	// release allocated ports
+	for _, port := range f.allocatedPorts {
 		f.portAllocator.Release(port)
 	}
-	f.allocedPorts = make([]int, 0)
+	f.allocatedPorts = make([]int, 0)
 
 	// clear os envs
 	f.osEnvs = make([]string, 0)
@@ -217,7 +217,7 @@ func (f *Framework) RenderTemplates(templates []string) (outs []string, ports ma
 	}
 
 	for _, t := range templates {
-		tmpl, err := template.New("").Parse(t)
+		tmpl, err := template.New("frp-e2e").Parse(t)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -237,7 +237,7 @@ func (f *Framework) PortByName(name string) int {
 func (f *Framework) AllocPort() int {
 	port := f.portAllocator.Get()
 	ExpectTrue(port > 0, "alloc port failed")
-	f.allocedPorts = append(f.allocedPorts, port)
+	f.allocatedPorts = append(f.allocatedPorts, port)
 	return port
 }
 
